requirements.txt
selenium
pillow
chromedriver-autoinstaller
pytest
allure-pytest
pytest-html
pytest-html-merger
beautifulsoup4
bs4